Sortbuffer

+++ src/clevel.h
@@ -5,6 +5,7 @@
 #include <filesystem>
 #include <atomic>
 #include "kvbuffer.h"
+#include "sortbuffer.h"
 #include "combotree_config.h"
 #include "debug.h"
 #include "pmem.h"
@@ -23,14 +24,13 @@ class __attribute__((packed)) CLevel {
 
  private:
   struct __attribute__((aligned(64))) Node {
-    enum class Type : uint8_t {
+    enum class Type : uint16_t {
       INVALID,
       INDEX,
       LEAF,
     };
 
     Type type;
-    uint8_t __padding[7];           // not used
     union {
       // uint48_t
       uint8_t next[6];              // used when type == LEAF. LSB == 1 means NULL
@@ -38,22 +38,21 @@ class __attribute__((packed)) CLevel {
     };
     union {
       // contains 2 bytes meta
-      KVBuffer<48+64,8> leaf_buf;   // used when type == LEAF
-      KVBuffer<48+64,6> index_buf;  // used when type == INDEX
+      SortBuffer<112,8> leaf_buf;   // used when type == LEAF
+      SortBuffer<112,6> index_buf;  // used when type == INDEX
     };
 
     Node* Put(MemControl* mem, uint64_t key, uint64_t value, Node* parent);
+    bool Update(MemControl* mem, uint64_t key, uint64_t value);
     bool Get(MemControl* mem, uint64_t key, uint64_t& value) const;
     bool Delete(MemControl* mem, uint64_t key, uint64_t* value);
-#ifndef BUF_SORT
-    void PutChild(MemControl* mem, void* key, const Node* child);
-#endif
+    void PutChild(MemControl* mem, uint64_t key, const Node* child);
 
     ALWAYS_INLINE Node* GetChild(int pos, uint64_t base_addr) const {
       if (pos == 0)
         return (Node*)(READ_SIX_BYTE(first_child) + base_addr);
       else
-        return (Node*)(index_buf.value(pos-1) + base_addr);
+        return (Node*)(index_buf.sort_value(pos-1) + base_addr);
     }
 
     ALWAYS_INLINE Node* GetNext(uint64_t base_addr) const {
@@ -100,6 +99,7 @@ class __attribute__((packed)) CLevel {
     MemControl(std::string pmem_file, size_t file_size)
       : pmem_file_(pmem_file+std::to_string(file_id_++))
     {
+#ifdef USE_LIBPMEM
       int is_pmem;
       std::filesystem::remove(pmem_file_);
       pmem_addr_ = pmem_map_file(pmem_file_.c_str(), file_size + 64,
@@ -119,12 +119,22 @@ class __attribute__((packed)) CLevel {
 
       cur_addr_ = base_addr_;
       end_addr_ = (uint8_t*)pmem_addr_ + mapped_len_;
+#else // libvmmalloc
+      pmem_file_ = "";
+      pmem_addr_ = nullptr;
+      base_addr_ = (uint64_t)new (std::align_val_t{64}) uint8_t[file_size];
+      assert((base_addr_ & 63) == 0);
+      cur_addr_  = base_addr_;
+      end_addr_  = (uint8_t*)base_addr_ + file_size;
+#endif
     }
 
     ~MemControl() {
-      if (pmem_addr_) {
+      if (!pmem_file_.empty() && pmem_addr_) {
         pmem_unmap(pmem_addr_, mapped_len_);
         std::filesystem::remove(pmem_file_);
+      } else {
+        delete (uint8_t*)base_addr_;
       }
     }
 
@@ -133,6 +143,7 @@ class __attribute__((packed)) CLevel {
       assert((uint8_t*)cur_addr_.load() + sizeof(CLevel::Node) < end_addr_);
       CLevel::Node* ret = (CLevel::Node*)cur_addr_.fetch_add(sizeof(CLevel::Node));
       ret->type = type;
+      ret->leaf_buf.header = 0x0123456789AB'0000UL;
       ret->leaf_buf.suffix_bytes = suffix_len;
       ret->leaf_buf.prefix_bytes = 8 - suffix_len;
       ret->leaf_buf.entries = 0;
@@ -169,24 +180,14 @@ class __attribute__((packed)) CLevel {
       while (cur_ != nullptr && cur_->leaf_buf.Empty())
         cur_ = (const Node*)cur_->GetNext(mem_->BaseAddr());
       if (cur_) {
-#ifdef BUF_SORT
         bool exist;
         idx_ = cur_->leaf_buf.FindLE(start_key, exist);
         idx_ = exist ? idx_ : idx_ + 1;
-#else
-        cur_->leaf_buf.GetSortedIndex(sorted_index_);
-        for (idx_ = 0; idx_ < cur_->leaf_buf.entries; ++idx_)
-          if (cur_->leaf_buf.key(sorted_index_[idx_], prefix_key) >= start_key)
-            break;
-#endif
         if (idx_ >= cur_->leaf_buf.entries) {
           do {
             cur_ = (const Node*)cur_->GetNext(mem_->BaseAddr());
           } while (cur_ != nullptr && cur_->leaf_buf.Empty());
           idx_ = 0;
-#ifndef BUF_SORT
-          if (cur_) cur_->leaf_buf.GetSortedIndex(sorted_index_);
-#endif
         }
       }
     }
@@ -198,25 +199,14 @@ class __attribute__((packed)) CLevel {
       while (cur_ != nullptr && cur_->leaf_buf.Empty())
         cur_ = (const Node*)cur_->GetNext(mem_->BaseAddr());
       idx_ = 0;
-#ifndef BUF_SORT
-      if (cur_) cur_->leaf_buf.GetSortedIndex(sorted_index_);
-#endif
     }
 
     ALWAYS_INLINE uint64_t key() const {
-#ifdef BUF_SORT
-      return cur_->leaf_buf.key(idx_, prefix_key);
-#else
-      return cur_->leaf_buf.key(sorted_index_[idx_], prefix_key);
-#endif
+      return cur_->leaf_buf.sort_key(idx_, prefix_key);
     }
 
     ALWAYS_INLINE uint64_t value() const {
-#ifdef BUF_SORT
-      return cur_->leaf_buf.value(idx_);
-#else
-      return cur_->leaf_buf.value(sorted_index_[idx_]);
-#endif
+      return cur_->leaf_buf.sort_value(idx_);
     }
 
     // return false if reachs end
@@ -227,9 +217,6 @@ class __attribute__((packed)) CLevel {
           cur_ = (const Node*)cur_->GetNext(mem_->BaseAddr());
         } while (cur_ != nullptr && cur_->leaf_buf.Empty());
         idx_ = 0;
-#ifndef BUF_SORT
-        if (cur_) cur_->leaf_buf.GetSortedIndex(sorted_index_);
-#endif
         return cur_ == nullptr ? false : true;
       } else {
         return true;
@@ -245,9 +232,6 @@ class __attribute__((packed)) CLevel {
     uint64_t prefix_key;
     const Node* cur_;
     int idx_;   // current index in node
-#ifndef BUF_SORT
-    int sorted_index_[16];
-#endif
   };
 
   class NoSortIter {
@@ -273,11 +257,11 @@ class __attribute__((packed)) CLevel {
     }
 
     ALWAYS_INLINE uint64_t key() const {
-      return cur_->leaf_buf.key(idx_, prefix_key);
+      return cur_->leaf_buf.sort_key(idx_, prefix_key);
     }
 
     ALWAYS_INLINE uint64_t value() const {
-      return cur_->leaf_buf.value(idx_);
+      return cur_->leaf_buf.sort_value(idx_);
     }
 
     // return false if reachs end
@@ -310,6 +294,10 @@ class __attribute__((packed)) CLevel {
   void Setup(MemControl* mem, int suffix_len);
   void Setup(MemControl* mem, KVBuffer<48+64,8>& buf);
   bool Put(MemControl* mem, uint64_t key, uint64_t value);
+
+  ALWAYS_INLINE bool Update(MemControl* mem, uint64_t key, uint64_t value) {
+    return root(mem->BaseAddr())->Update(mem, key, value);
+  }
 
   ALWAYS_INLINE bool Get(MemControl* mem, uint64_t key, uint64_t& value) const {
     return root(mem->BaseAddr())->Get(mem, key, value);
